refactor: extract xwayland.c from somewm.c
Move all XWayland X11 compatibility handlers into their own module:
activatex11, associatex11, configurex11, createnotifyx11,
dissociatex11, sethints, xwaylandready.

Add xwayland_setup() and xwayland_cleanup() to encapsulate XWayland
lifecycle, replacing inline code in setup() and cleanup().

Make mapnotify, unmapnotify, destroynotify, fullscreennotify,
updatetitle, and get_border_width non-static since they are called
from xwayland.c (and will be needed by window.c later).

somewm.c: 7062 -> 6799 lines (-263)
xwayland.c: 311 lines (new)
